refactor: move ProfileButton to /src/components/user and update imports
